agregamos funciones del 7 al 13 y actualizamos

+++ modelo/NumeroModel.cs
@@ -2,7 +2,7 @@ namespace modelo
 {
     public class NumeroModel
     {   // declarando las variables
-        public decimal Numero_A { get; set; }
-        public decimal Numero_B { get; set; }
+        public double Numero_A { get; set; }
+        public double Numero_B { get; set; }
     }
 }
